move all except app.js to src/
